+++ app/src/main/java/io/iamjosephmj/flingersample/ui/utils/NavigationSetup.kt
@@ -31,7 +31,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import io.iamjosephmj.flingersample.ui.comparison.ComparisonScreen
 import io.iamjosephmj.flingersample.ui.demos.AdaptiveDemo
-import io.iamjosephmj.flingersample.ui.demos.DebugOverlayDemo
 import io.iamjosephmj.flingersample.ui.demos.PagerDemo
 import io.iamjosephmj.flingersample.ui.demos.SnapGalleryDemo
 import io.iamjosephmj.flingersample.ui.home.HomeScreen
@@ -49,7 +48,6 @@ import io.iamjosephmj.flingersample.ui.presets.PresetsGalleryScreen
  * - "snapDemo" - Snap-to-item gallery demo
  * - "adaptiveDemo" - Velocity-aware adaptive fling demo
  * - "pagerDemo" - Custom pager physics demo
- * - "debugDemo" - Debug overlay visualization demo
  */
 @Composable
 fun CreateNavHost(navController: NavHostController) {
@@ -67,6 +65,5 @@ fun CreateNavHost(navController: NavHostController) {
         composable("snapDemo") { SnapGalleryDemo(navController) }
         composable("adaptiveDemo") { AdaptiveDemo(navController) }
         composable("pagerDemo") { PagerDemo(navController) }
-        composable("debugDemo") { DebugOverlayDemo(navController) }
     }
 }